requirements.txt
Pillow
PyQt6
Matplotlib
pymupdf